reformat solution 3D1P AoC2020
I liked making the mountain grow large with the first answer, but thought that if it got too large it would freeze the computer. also it is hard to follow with the eyes.  this update was just a modification to the previous answer that makes it easier to follow and  i suspect less intensive on the computer